# BookManagementSystem/swagger.json
{
  "openapi": "3.1.0",
  "info": {
      "title": "BookManagement",
      "description": "API documentation for Book Management system",
      "version": "1.0.0",
      "contact": {}
  },
  "servers": [
      {
          "url": "http://localhost:3000",
          "description" : "Local Server"
      },
      {
        "url": "http://localhost:4000",
        "description" : "Production Server"
      }
  ],
  "paths": {
      "/api/users/signup": {
          "post": {
              "tags": [
                  "User"
              ],
              "summary": "Signup",
              "description": "Signup",
              "operationId": "signup",
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "type": "object",
                              "properties": {
                                  "email": {
                                      "type": "string",
                                      "example": "mitaligadhiya6@gmail.com"
                                  },
                                  "password": {
                                      "type": "string",
                                      "example": "1234"
                                  },
                                  "role": {
                                      "type": "string",
                                      "example": "admin"
                                  },
                                  "username": {
                                      "type": "string",
                                      "example": "Mitali"
                                  }
                              }
                          },
                          "examples": {
                              "Signup": {
                                  "value": {
                                      "email": "mitaligadhiya6@gmail.com",
                                      "password": "1234",
                                      "role": "admin",
                                      "username": "Mitali"
                                  }
                              }
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "Ok Signed Up succesfully"
                  },
                  "500" : {
                    "description": "Internal Server Error"
                  }
              }
          }
      },
      "/api/users/login": {
          "post": {
              "tags": [
                  "User"
              ],
              "summary": "Login",
              "description": "Login",
              "operationId": "login",
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "type": "object",
                              "properties": {
                                  "email": {
                                      "type": "string",
                                      "example": "mitaligadhiya6@gmail.com"
                                  },
                                  "password": {
                                      "type": "string",
                                      "example": "1234"
                                  }
                              }
                          },
                          "examples": {
                              "Login": {
                                  "value": {
                                      "email": "mitaligadhiya6@gmail.com",
                                      "password": "1234"
                                  }
                              }
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": ""
                  }
              }
          }
      },
      "/api/users/protected": {
          "get": {
              "tags": [
                  "User"
              ],
              "summary": "Protected",
              "description": "Protected",
              "operationId": "protected",
              "responses": {
                  "200": {
                      "description": ""
                  }
              },
              "security": [
                  {
                      "bearerAuth": []
                  }
              ]
          }
      },
...
  },
  "components": {
      "securitySchemes": {
          "bearerAuth": {
              "type": "http",
              "scheme": "bearer"
          }
        },
...
  },
...
}